requirements.txt
ArchivesSnake==0.8.2
attrs==19.3.0
boltons==19.3.0
certifi==2019.9.11
chardet==3.0.4
clinner==1.12.3
colorlog==4.0.2
Django==2.2.7
django-cors-headers==3.2.0
django-js-asset==1.2.2
django-mptt==0.10.0
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
gitdb2==2.0.6
GitPython==3.0.5
health-check==3.4.1
idna==2.8
more-itertools==7.2.0
psycopg2-binary==2.8.4
PyJWT==1.7.1
pytz==2019.3
PyYAML==5.1.2
requests==2.22.0
six==1.13.0
smmap2==2.0.5
sqlparse==0.3.0
structlog==19.2.0
uritemplate==3.0.0
urllib3==1.25.7
